WIP: Removed old CSP implementation and updated CSP tests with wiremock verify
